@@ -20,36 +20,37 @@
 		"toggle": "tsx turbo/scripts/toggle.mts"
 	},
 	"dependencies": {
-		"@types/react": "^19.1.0"
+		"@types/react": "^19.1.2"
 	},
 	"devDependencies": {
 		"@commitlint/cli": "^19.8.0",
 		"@commitlint/config-conventional": "^19.8.0",
 		"@eslint/config-inspector": "^1.0.2",
-		"@eslint/js": "^9.24.0",
-		"@types/node": "^22.14.0",
-		"@typescript-eslint/utils": "^8.29.1",
+		"@eslint/js": "^9.25.0",
+		"@types/node": "^22.14.1",
+		"@typescript-eslint/utils": "^8.30.1",
 		"@vitest/coverage-istanbul": "^3.1.1",
 		"@vitest/coverage-v8": "^3.1.1",
 		"@vitest/ui": "^3.1.1",
 		"comment-json": "^4.2.5",
-		"eslint": "^9.24.0",
-		"eslint-config-prettier": "^10.1.1",
-		"eslint-import-resolver-typescript": "^4.3.2",
-		"eslint-plugin-import-x": "^4.10.2",
-		"eslint-plugin-package-json": "^0.29.0",
+		"eslint": "^9.25.0",
+		"eslint-config-prettier": "^10.1.2",
+		"eslint-import-resolver-typescript": "^4.3.3",
+		"eslint-plugin-import-x": "^4.10.6",
+		"eslint-plugin-package-json": "^0.29.1",
 		"eslint-plugin-prettier": "^5.2.6",
 		"globals": "^16.0.0",
 		"husky": "^9.1.7",
 		"jiti": "^2.4.2",
 		"prettier": "^3.5.3",
 		"prettier-plugin-packagejson": "^2.5.10",
+		"ts-dedent": "^2.2.0",
 		"tsup": "^8.4.0",
 		"tsx": "^4.19.3",
 		"turbo": "^2.5.0",
-		"type-fest": "^4.39.1",
+		"type-fest": "^4.40.0",
 		"typescript": "^5.8.3",
-		"typescript-eslint": "^8.29.1",
+		"typescript-eslint": "^8.30.1",
 		"vitest": "^3.1.1"
 	},
 	"packageManager": "pnpm@10.7.1"
