@@ -1,54 +1,22 @@
-import { readFileSync } from "node:fs";
-import { dirname } from "node:path";
-import { AST_NODE_TYPES, ESLintUtils } from "@typescript-eslint/utils";
-import { SymbolFlags } from "typescript";
-import {
-	serverSafeHooks,
-	clientOnlyHooks,
-	browserAPIs,
-	loadModuleCategories,
-	isClientDetectionCondition,
-	isHookCall,
-	getFileAST,
-	clientOnlyPackages,
-	clientSideDependencyCache,
-	hasFileChanged,
-	mightBeCompiledClientCode
-} from "../utils.js";
-import type { TSESLint, TSESTree } from "@typescript-eslint/utils";
+import { ESLintUtils, AST_NODE_TYPES } from "@typescript-eslint/utils";
+import type { TSESTree } from "@typescript-eslint/utils";
 
-type MessageIds = "missingUseClient" | "detectedClientDep" | "sharedComponent";
+type MessageIds = "missingUseClient";
+type RuleOptions = [{ allowlist: Record<string, boolean | string[]> }];
 
-// Define the types for the rule options
-export type AllowlistType = Record<string, boolean | string[]>;
+const DOM_APIS = new Set(["document", "window", "navigator", "localStorage", "sessionStorage", "history", "location"]);
 
-export interface RuleOptionsObject {
-	allowlist: AllowlistType;
-	traceDepth: number;
-	traceDependencies: boolean;
-}
-
-export type RuleOptions = [RuleOptionsObject];
-
-const createRule = ESLintUtils.RuleCreator<{
-	recommended: boolean;
-}>((name) => `https://spencerbeg.gs/rule/${name}`);
-
-export const enforceRule = createRule<RuleOptions, MessageIds>({
+export const useClientDirectiveRule = ESLintUtils.RuleCreator((name) => `https://spencerbeg.gs/rule/${name}`)<
+	RuleOptions,
+	MessageIds
+>({
 	name: "use-client-directive",
 	meta: {
 		type: "problem",
-		fixable: "code",
 		docs: {
-			description: 'Enforces the use of "use client" directive in components with client-side dependencies',
-			recommended: true
+			description: "Enforce 'use client' directive when using DOM APIs"
 		},
-		messages: {
-			missingUseClient: 'Component with client-side dependencies is missing "use client" directive',
-			detectedClientDep: "Detected client-side dependency: {{ dependency }}",
-			sharedComponent:
-				"This appears to be a shared component. Consider creating separate client/server versions or using dynamic imports"
-		},
+		fixable: "code",
 		schema: [
 			{
 				type: "object",
@@ -64,949 +32,224 @@ export const enforceRule = createRule<RuleOptions, MessageIds>({
 								}
 							]
 						}
-					},
-					traceDepth: {
-						type: "integer",
-						minimum: 0
-					},
-					traceDependencies: {
-						type: "boolean"
 					}
 				},
 				additionalProperties: false
 			}
-		]
+		],
+		messages: {
+			missingUseClient: "Component uses DOM API '{{ api }}' and must be marked with 'use client' directive"
+		}
 	},
 	defaultOptions: [
 		{
-			allowlist: {
-				react: true,
-				"react-dom": true,
-				next: true,
-				"next/image": true,
-				"next/link": true,
-				"next/router": true
-			},
-			traceDepth: 1,
-			traceDependencies: true
+			allowlist: {}
 		}
 	],
-	create(context): TSESLint.RuleListener {
-		// Skip analysis if no type information available
-		const services = context.sourceCode.parserServices;
+	create(context, options) {
+		const allowedModules = options[0].allowlist;
+		const sourceCode = context.sourceCode;
+		const reportedNodes = new Set<string>();
+		const imports = new Map<string, { source: string; isNamespaceImport: boolean; importedName: string }>();
+		const domAliases = new Map<string, string>(); // Maps variable names to their DOM API names
+		let hasUseClientDirective = false;
 
-		if (!services?.program) {
-			context.report({
-				messageId: "detectedClientDep",
-				data: { dependency: "No type information available" },
-				node: context.sourceCode.ast
-			});
-			return {};
-		}
-
-		// Use the internal factory to create visitors with the context
-		// Force non-null assertion since we've checked for program existence
-		return createRuleVisitors(context, {
-			program: services.program,
-			esTreeNodeToTSNodeMap: services.esTreeNodeToTSNodeMap
-		});
-	}
-});
-
-// Define a safer interface for the services parameter
-interface TypeInformation {
-	program: {
-		getTypeChecker: () => unknown;
-	};
-	esTreeNodeToTSNodeMap?: unknown;
-}
-
-// Expose a factory function for creating visitor functions
-// This makes the rule's logic more testable by separating the context validation
-// from the actual rule logic
-export function createRuleVisitors(
-	context: Readonly<TSESLint.RuleContext<MessageIds, RuleOptions>>,
-	services: TypeInformation
-): TSESLint.RuleListener {
-	const options = context.options[0];
-	const { allowlist, traceDepth, traceDependencies } = options;
-
-	const checker = services.program.getTypeChecker();
-	const fileName = context.filename;
-	const sourceCode = context.sourceCode;
-
-	// Try to load module categories from config file
-	const moduleCategories = loadModuleCategories();
-
-	// Extend with configured hooks
-	const serverSafeHooksExtended = [...serverSafeHooks, ...moduleCategories.serverSafeHooks];
-	const clientOnlyHooksExtended = [...clientOnlyHooks, ...moduleCategories.clientOnlyHooks];
-
-	// Track if 'use client' directive exists in the file
-	let hasUseClientDirective = false;
-
-	// Track detected client dependencies
-	const clientDependencies: string[] = [];
-
-	// Track current trace depth for dependency resolution
-	let currentTraceDepth = 0;
-
-	// Set of files checked in this run to avoid circular dependencies
-	const filesCheckedInThisRun = new Set<string>();
-
-	// Track exports to detect shared components
-	let hasExports = false;
-	let hasDirectClientAPIs = false;
-
-	// Check if a module is allowlisted
-	function isModuleAllowlisted(moduleName: string): boolean {
-		if (allowlist[moduleName] === true) {
-			return true;
-		}
-
-		// Check against loaded module categories
-		if (moduleCategories.serverModules.includes(moduleName)) {
-			return true;
-		}
-
-		return false;
-	}
-
-	// Check if a specific export from a module is allowlisted
-	function isExportAllowlisted(exportName: string, moduleName: string, allowlist: AllowlistType): boolean {
-		const allowlistValue = allowlist[moduleName];
-
-		if (allowlistValue === true) {
-			return true;
-		}
-
-		if (Array.isArray(allowlistValue)) {
-			return allowlistValue.includes(exportName);
-		}
-
-		return false;
-	}
-
-	// Check if a package is known to be client-only
-	function isKnownClientOnlyPackage(packageName: string): boolean {
-		// Check against configured client modules
-		if (moduleCategories.clientModules.includes(packageName)) {
-			return true;
-		}
-
-		// Check against hardcoded list
-		return clientOnlyPackages.some((prefix) => packageName === prefix || packageName.startsWith(prefix));
-	}
-
-	// Function to check if a node uses client-side APIs
-	function checkNodeForClientAPIs(node: TSESTree.Node): boolean {
-		// Check for React hook usage
-		if (isHookCall(node)) {
-			const callExpr = node as TSESTree.CallExpression;
-			const callee = callExpr.callee;
-			const hookName = callee.type === AST_NODE_TYPES.Identifier ? callee.name : "";
-
-			// Check if it's a client-only hook
-			if (clientOnlyHooksExtended.includes(hookName)) {
-				clientDependencies.push(`React hook: ${hookName}`);
-				hasDirectClientAPIs = true;
+		function isAllowlistedImport(source: string, importedName: string): boolean {
+			// If the module is not in the allowlist, allow it by default
+			if (!(source in allowedModules)) {
 				return true;
 			}
 
-			// If it's a server-safe hook, we can ignore it
-			if (serverSafeHooksExtended.includes(hookName)) {
-				return false;
+			const allowedExports = allowedModules[source];
+			if (allowedExports === true) {
+				return true;
 			}
-
-			// Unknown hooks should be treated conservatively
-			clientDependencies.push(`Unknown hook call: ${hookName}`);
-			hasDirectClientAPIs = true;
-			return true;
+			if (Array.isArray(allowedExports)) {
+				return allowedExports.includes(importedName) || allowedExports.includes("*");
+			}
+			return false;
 		}
 
-		// Check for DOM globals
-		if (node.type === AST_NODE_TYPES.Identifier) {
-			const name = node.name;
+		function reportNode(node: TSESTree.Node, api: string): void {
+			if (hasUseClientDirective) {
+				return;
+			}
 
-			if (browserAPIs.includes(name)) {
-				// Get the TS node to verify it's a global
-				const tsNodeMap = services.esTreeNodeToTSNodeMap as { get?: (node: TSESTree.Node) => unknown };
-				const tsNode = tsNodeMap.get?.(node);
+			// For member expressions, use the object's location
+			const loc = node.type === AST_NODE_TYPES.MemberExpression ? node.object.loc : node.loc;
+			const nodeKey = `${String(loc.start.line)}:${String(loc.start.column)}:${api}`;
 
-				// We need to check symbol flags, so ensure we have a checker and node
-				if (tsNode) {
-					try {
-						// Use type assertion to access getSymbolAtLocation with better error handling
-						const getSymbolFn = (
-							checker as {
-								getSymbolAtLocation: (node: unknown) => { flags: number } | undefined;
-							}
-						).getSymbolAtLocation;
+			if (!reportedNodes.has(nodeKey)) {
+				reportedNodes.add(nodeKey);
+				context.report({
+					node,
+					messageId: "missingUseClient",
+					data: { api },
+					fix: (fixer) => {
+						// Add 'use client' directive at the beginning of the file
+						const comments = sourceCode.getAllComments();
+						const leadingComments = comments.filter((comment) => comment.loc.start.line === 1);
 
-						const symbol = getSymbolFn(tsNode);
-
-						// Check if it's actually the global object and not a local variable
-						if (symbol && (symbol.flags & SymbolFlags.Value) !== 0) {
-							clientDependencies.push(`Browser API: ${name}`);
-							hasDirectClientAPIs = true;
-							return true;
+						if (leadingComments.length > 0) {
+							// If there are leading comments, add the directive after them
+							const lastLeadingComment = leadingComments[leadingComments.length - 1];
+							return fixer.insertTextAfter(lastLeadingComment, "\n'use client';");
+						} else {
+							// Otherwise, add it at the beginning of the file
+							return fixer.insertTextBefore(sourceCode.ast, "'use client';\n");
 						}
-					} catch (_error) {
-						// If any error occurs during type checking, log and continue
 					}
-				}
+				});
 			}
 		}
 
-		// Check for window/document property access
-		if (node.type === AST_NODE_TYPES.MemberExpression) {
+		function checkMemberExpression(node: TSESTree.MemberExpression): void {
 			if (node.object.type === AST_NODE_TYPES.Identifier) {
 				const objectName = node.object.name;
-				if (browserAPIs.includes(objectName)) {
-					const propName = node.property.type === AST_NODE_TYPES.Identifier ? node.property.name : "?";
-					clientDependencies.push(`Browser API: ${objectName}.${propName}`);
-					hasDirectClientAPIs = true;
-					return true;
-				}
-			}
-		}
+				const importInfo = imports.get(objectName);
+				const propertyName = node.property.type === AST_NODE_TYPES.Identifier ? node.property.name : undefined;
 
-		// Check for event handlers in JSX attributes
-		if (node.type === AST_NODE_TYPES.JSXAttribute) {
-			const attrName = node.name.name;
-			if (typeof attrName === "string" && attrName.startsWith("on") && attrName.length > 2) {
-				// Check if the second character is uppercase (like onClick, onChange)
-				const secondChar = attrName.charAt(2);
-				if (secondChar === secondChar.toUpperCase()) {
-					clientDependencies.push(`JSX event handler: ${attrName}`);
-					hasDirectClientAPIs = true;
-					return true;
-				}
-			}
-		}
-
-		// Check if statements for client detection pattern
-		if (node.type === AST_NODE_TYPES.IfStatement) {
-			if (isClientDetectionCondition(node.test)) {
-				clientDependencies.push("Client detection condition (typeof window check)");
-				hasDirectClientAPIs = true;
-				return true;
-			}
-		}
-
-		return false;
-	}
-
-	// Function to check imports for client dependencies
-	function checkImportForClientDependencies(node: TSESTree.Node): boolean {
-		if (node.type !== AST_NODE_TYPES.ImportDeclaration) {
-			return false;
-		}
-
-		const moduleSpecifier = node.source.value;
-
-		// Skip completely allowlisted modules
-		if (options.allowlist[moduleSpecifier] === true) {
-			return false;
-		}
-
-		// Check if it's a known client-only package
-		if (isKnownClientOnlyPackage(moduleSpecifier)) {
-			clientDependencies.push(`Import from client-only package: ${moduleSpecifier}`);
-			hasDirectClientAPIs = true;
-			return true;
-		}
-
-		// Skip type-only imports
-		if (node.importKind === "type") {
-			return false;
-		}
-
-		// Check if any specific named imports are allowlisted
-		let hasClientDependency = false;
-
-		for (const specifier of node.specifiers) {
-			// Skip type-only imports
-			if (specifier.type === AST_NODE_TYPES.ImportSpecifier && specifier.importKind === "type") {
-				continue;
-			}
-
-			if (specifier.type === AST_NODE_TYPES.ImportSpecifier) {
-				const importedName = specifier.imported.type === AST_NODE_TYPES.Identifier ? specifier.imported.name : "";
-
-				// If this specific import is allowlisted, skip it
-				if (isExportAllowlisted(importedName, moduleSpecifier, allowlist)) {
-					continue;
-				}
-
-				// Trace dependencies if needed and within depth limit
-				if (traceDependencies && traceDepth > currentTraceDepth) {
-					try {
-						// Resolve the module path
-						const resolvedPath = resolveModulePath(moduleSpecifier);
-
-						// Skip node_modules or ensure path exists
-						if (resolvedPath && !isNodeModule(resolvedPath)) {
-							// Check if this module is in our cache and hasn't been checked in this run
-							if (!filesCheckedInThisRun.has(resolvedPath)) {
-								if (clientSideDependencyCache.has(resolvedPath) && !hasFileChanged(resolvedPath)) {
-									if (clientSideDependencyCache.get(resolvedPath)) {
-										clientDependencies.push(
-											`Import from client module: ${moduleSpecifier}.${importedName}`
-										);
-										hasClientDependency = true;
-									}
-								} else {
-									// Analyze the dependency if not cached
-									const depHasClientSide = analyzeModuleForClientDependencies(resolvedPath);
-									if (depHasClientSide) {
-										clientDependencies.push(
-											`Import from client module: ${moduleSpecifier}.${importedName}`
-										);
-										hasClientDependency = true;
-									}
-								}
-							}
+				if (importInfo && propertyName) {
+					if (importInfo.isNamespaceImport) {
+						// For namespace imports, first check if the entire namespace is allowlisted
+						if (isAllowlistedImport(importInfo.source, "*")) {
+							return;
 						}
-					} catch (_err: unknown) {
-						// Continue with other attempts
-					}
-				}
-			}
-		}
 
-		return hasClientDependency;
-	}
+						// Then check if the specific property is allowlisted
+						if (isAllowlistedImport(importInfo.source, propertyName)) {
+							return;
+						}
 
-	// Function to check dynamic imports
-	function checkDynamicImport(node: TSESTree.Node): boolean {
-		if (node.type !== AST_NODE_TYPES.ImportExpression) {
-			return false;
-		}
-
-		// Check the source of the dynamic import
-		if (node.source.type === AST_NODE_TYPES.Literal && typeof node.source.value === "string") {
-			const moduleSpecifier = node.source.value;
-
-			// Skip allowlisted modules
-			if (isModuleAllowlisted(moduleSpecifier)) {
-				return false;
-			}
-
-			// Check if it's a known client-only package
-			if (isKnownClientOnlyPackage(moduleSpecifier)) {
-				clientDependencies.push(`Dynamic import from client-only package: ${moduleSpecifier}`);
-				hasDirectClientAPIs = true;
-				return true;
-			}
-
-			// Trace dependencies if needed and within depth limit
-			if (traceDependencies && traceDepth > currentTraceDepth) {
-				try {
-					// Resolve the module path
-					const resolvedPath = resolveModulePath(moduleSpecifier);
-
-					// Skip node_modules or ensure path exists
-					if (resolvedPath && !isNodeModule(resolvedPath)) {
-						// Check if this module is in our cache
-						if (!filesCheckedInThisRun.has(resolvedPath)) {
-							if (clientSideDependencyCache.has(resolvedPath) && !hasFileChanged(resolvedPath)) {
-								if (clientSideDependencyCache.get(resolvedPath)) {
-									clientDependencies.push(`Dynamic import from client module: ${moduleSpecifier}`);
-									hasDirectClientAPIs = true;
-									return true;
-								}
-							} else {
-								// Analyze the dependency if not cached
-								const depHasClientSide = analyzeModuleForClientDependencies(resolvedPath);
-								if (depHasClientSide) {
-									clientDependencies.push(`Dynamic import from client module: ${moduleSpecifier}`);
-									hasDirectClientAPIs = true;
-									return true;
-								}
+						// For non-allowlisted namespace imports, report if it's a DOM API
+						if (DOM_APIS.has(propertyName) && importInfo.source in allowedModules) {
+							reportNode(node, propertyName);
+						}
+					} else {
+						// For named imports, check if the import is allowlisted
+						if (!isAllowlistedImport(importInfo.source, importInfo.importedName)) {
+							if (DOM_APIS.has(importInfo.importedName)) {
+								reportNode(node, importInfo.importedName);
 							}
 						}
 					}
-				} catch (_err: unknown) {
-					// Continue with other attempts
+				} else if (DOM_APIS.has(objectName)) {
+					reportNode(node.object, objectName);
 				}
+			} else if (node.object.type === AST_NODE_TYPES.MemberExpression) {
+				checkMemberExpression(node.object);
 			}
 		}
 
-		return false;
-	}
-
-	// Helper to resolve a module path
-	function resolveModulePath(moduleSpecifier: string): string | null {
-		try {
-			// Try to resolve relative to the current file
-			const resolvedPath = require.resolve(moduleSpecifier, {
-				paths: [dirname(fileName)]
-			});
-			return resolvedPath;
-		} catch (_err: unknown) {
-			// Module resolution failed
-			return null;
-		}
-	}
-
-	// Helper to check if a path is in node_modules
-	function isNodeModule(path: string): boolean {
-		return path.includes("node_modules");
-	}
-
-	// Helper to analyze a module for client dependencies
-	function analyzeModuleForClientDependencies(filePath: string): boolean {
-		// Increment trace depth
-		currentTraceDepth++;
-
-		// Add to set of checked files
-		filesCheckedInThisRun.add(filePath);
-
-		try {
-			// Get AST for the file
-			const result = getFileAST(filePath);
-
-			// If we have a cached result, use it
-			if (result.hasClientCode !== undefined) {
-				return result.hasClientCode;
+		function checkIdentifier(node: TSESTree.Identifier): void {
+			// Skip identifiers that are part of import/export declarations
+			if (
+				node.parent.type === AST_NODE_TYPES.ImportSpecifier ||
+				node.parent.type === AST_NODE_TYPES.ImportDefaultSpecifier ||
+				node.parent.type === AST_NODE_TYPES.ImportNamespaceSpecifier ||
+				node.parent.type === AST_NODE_TYPES.ExportSpecifier
+			) {
+				return;
 			}
 
-			// If there was an error analyzing the file, assume it's client-side
-			if (result.error) {
-				return true;
+			// Skip identifiers that are part of MemberExpressions as the object
+			if (node.parent.type === AST_NODE_TYPES.MemberExpression && node.parent.object === node) {
+				return;
 			}
 
-			// If the file has 'use client', it's definitely client-side
-			if (result.hasUseClient) {
-				return true;
+			// Skip identifiers that are properties of MemberExpressions
+			if (node.parent.type === AST_NODE_TYPES.MemberExpression && node.parent.property === node) {
+				return;
 			}
 
-			// Otherwise, check if the content might be compiled client code
-			const content = readFileSync(filePath, "utf8");
-			return mightBeCompiledClientCode(content);
-		} catch (_err: unknown) {
-			// If we can't analyze the file, assume it's client-side
-			return true;
-		} finally {
-			// Decrement trace depth
-			currentTraceDepth--;
-		}
-	}
-
-	return {
-		Program(node) {
-			// Reset tracking state
-			hasUseClientDirective = false;
-			clientDependencies.length = 0;
-			hasExports = false;
-			hasDirectClientAPIs = false;
-			filesCheckedInThisRun.clear();
-
-			// Check for 'use client' directive
-			const comments = sourceCode.getAllComments();
-			hasUseClientDirective = comments.some((comment) => comment.value.trim() === "use client");
-
-			// Check for client dependencies in the program
-			for (const statement of node.body) {
-				if (checkNodeForClientAPIs(statement)) {
-					hasDirectClientAPIs = true;
-				}
+			// Skip identifiers that are part of variable declarations
+			if (node.parent.type === AST_NODE_TYPES.VariableDeclarator && node.parent.id === node) {
+				return;
 			}
 
-			// Force a call to isHookCall for testing purposes
-			const _dummyNode = {
-				type: AST_NODE_TYPES.CallExpression,
-				callee: {
-					type: AST_NODE_TYPES.Identifier,
-					name: "useState"
-				}
-			} as TSESTree.CallExpression;
-			isHookCall(_dummyNode);
-		},
-
-		CallExpression(node) {
-			// Check for hook calls
-			if (isHookCall(node)) {
-				checkNodeForClientAPIs(node);
-			} else if (node.callee.type === AST_NODE_TYPES.Identifier && node.callee.name.startsWith("use")) {
-			}
-		},
-
-		VariableDeclarator(node) {
-			// Check for hook calls in variable declarations
-			if (node.init && node.init.type === AST_NODE_TYPES.CallExpression) {
-				if (isHookCall(node.init)) {
-					checkNodeForClientAPIs(node.init);
-				}
-			}
-		},
-
-		ExpressionStatement(node) {
-			// Check for hook calls in expressions
-			if (node.expression.type === AST_NODE_TYPES.CallExpression) {
-				if (isHookCall(node.expression)) {
-					checkNodeForClientAPIs(node.expression);
-				}
-			}
-		},
-
-		ImportDeclaration(node) {
-			checkImportForClientDependencies(node);
-		},
-
-		ImportExpression(node) {
-			checkDynamicImport(node);
-		},
-
-		ExportNamedDeclaration() {
-			hasExports = true;
-		},
-
-		ExportDefaultDeclaration() {
-			hasExports = true;
-		},
-
-		"Program:exit"() {
-			// If we have client dependencies but no 'use client' directive
-			if (clientDependencies.length > 0 && !hasUseClientDirective) {
-				context.report({
-					messageId: "missingUseClient",
-					node: context.sourceCode.ast,
-					fix(fixer) {
-						return fixer.insertTextBefore(context.sourceCode.ast, "'use client';\n\n");
+			const importInfo = imports.get(node.name);
+			if (importInfo) {
+				if (!isAllowlistedImport(importInfo.source, importInfo.importedName)) {
+					if (DOM_APIS.has(importInfo.importedName)) {
+						reportNode(node, importInfo.importedName);
 					}
+				}
+			} else if (DOM_APIS.has(node.name)) {
+				reportNode(node, node.name);
+			}
+		}
+
+		return {
+			Program(node: TSESTree.Program): void {
+				// Check for 'use client' directive
+				hasUseClientDirective = node.body.some((statement) => {
+					if (statement.type === AST_NODE_TYPES.ExpressionStatement) {
+						const expression = statement.expression;
+						if (expression.type === AST_NODE_TYPES.Literal && typeof expression.value === "string") {
+							return expression.value.trim() === "use client";
+						}
+					}
+					return false;
 				});
 
-				// Report each detected dependency
-				for (const dep of clientDependencies) {
-					context.report({
-						messageId: "detectedClientDep",
-						data: { dependency: dep },
-						node: context.sourceCode.ast
+				// Also check comments for 'use client' directive
+				if (!hasUseClientDirective) {
+					hasUseClientDirective = sourceCode.getAllComments().some((comment) => {
+						return comment.value.trim() === "use client";
 					});
 				}
-			}
 
-			// If we have exports but no direct client APIs, suggest splitting into client/server components
-			if (hasExports && !hasDirectClientAPIs && clientDependencies.length > 0) {
-				context.report({
-					messageId: "sharedComponent",
-					node: context.sourceCode.ast
+				if (hasUseClientDirective) {
+					return;
+				}
+
+				// Track imports
+				node.body.forEach((statement) => {
+					if (statement.type === AST_NODE_TYPES.ImportDeclaration) {
+						const source = statement.source.value;
+						statement.specifiers.forEach((specifier) => {
+							if (specifier.type === AST_NODE_TYPES.ImportSpecifier) {
+								const importedName =
+									specifier.imported.type === AST_NODE_TYPES.Identifier
+										? specifier.imported.name
+										: specifier.imported.value;
+								imports.set(specifier.local.name, {
+									source,
+									importedName,
+									isNamespaceImport: false
+								});
+							} else if (specifier.type === AST_NODE_TYPES.ImportNamespaceSpecifier) {
+								imports.set(specifier.local.name, {
+									source,
+									importedName: "*",
+									isNamespaceImport: true
+								});
+							}
+						});
+					}
 				});
-			}
-		}
-	};
-}
+			},
 
-// Export internal functions for testing
-export function isModuleAllowlisted(moduleName: string, allowlist: AllowlistType): boolean {
-	if (allowlist[moduleName] === true) {
-		return true;
-	}
-
-	return false;
-}
-
-export function isExportAllowlisted(exportName: string, moduleName: string, allowlist: AllowlistType): boolean {
-	const allowlistValue = allowlist[moduleName];
-
-	if (allowlistValue === true) {
-		return true;
-	}
-
-	if (Array.isArray(allowlistValue)) {
-		return allowlistValue.includes(exportName);
-	}
-
-	return false;
-}
-
-// Export for testing
-export function isKnownClientOnlyPackage(
-	packageName: string,
-	moduleCategories: { clientModules: string[] } = { clientModules: [] }
-): boolean {
-	// Check against configured client modules
-	if (moduleCategories.clientModules.includes(packageName)) {
-		return true;
-	}
-
-	// Check against hardcoded list
-	return clientOnlyPackages.some((prefix) => packageName === prefix || packageName.startsWith(prefix));
-}
-
-// Export for testing - simplified version without context object
-export function checkForClientAPIs(
-	node: TSESTree.Node,
-	options?: {
-		clientOnlyHooksExtended?: string[];
-		serverSafeHooksExtended?: string[];
-		browserAPIs?: string[];
-	}
-): { isClient: boolean; dependency?: string } {
-	// Set default values if options is undefined
-	const clientOnlyHooksExtended = options?.clientOnlyHooksExtended ?? clientOnlyHooks;
-	const serverSafeHooksExtended = options?.serverSafeHooksExtended ?? serverSafeHooks;
-	const browserAPIList = options?.browserAPIs ?? browserAPIs;
-
-	// Check for React hook usage
-	if (isHookCall(node)) {
-		const callExpr = node as TSESTree.CallExpression;
-		const callee = callExpr.callee;
-		const hookName = callee.type === AST_NODE_TYPES.Identifier ? callee.name : "";
-
-		// Check if it's a client-only hook
-		if (clientOnlyHooksExtended.includes(hookName)) {
-			return { isClient: true, dependency: `React hook: ${hookName}` };
-		}
-
-		// If it's a server-safe hook, we can ignore it
-		if (serverSafeHooksExtended.includes(hookName)) {
-			return { isClient: false };
-		}
-
-		// Unknown hooks should be treated conservatively
-		return { isClient: true, dependency: `Unknown hook call: ${hookName}` };
-	}
-
-	// Check for DOM globals
-	if (node.type === AST_NODE_TYPES.Identifier) {
-		const name = node.name;
-
-		if (browserAPIList.includes(name)) {
-			return { isClient: true, dependency: `Browser API: ${name}` };
-		}
-	}
-
-	// Check for window/document property access
-	if (node.type === AST_NODE_TYPES.MemberExpression) {
-		if (node.object.type === AST_NODE_TYPES.Identifier) {
-			const objectName = node.object.name;
-			if (browserAPIList.includes(objectName)) {
-				const propName = node.property.type === AST_NODE_TYPES.Identifier ? node.property.name : "?";
-				return { isClient: true, dependency: `Browser API: ${objectName}.${propName}` };
-			}
-		}
-	}
-
-	// Check for event handlers in JSX attributes
-	if (node.type === AST_NODE_TYPES.JSXAttribute) {
-		const attrName = node.name.name;
-		if (typeof attrName === "string" && attrName.startsWith("on") && attrName.length > 2) {
-			// Check if the second character is uppercase (like onClick, onChange)
-			const secondChar = attrName.charAt(2);
-			if (secondChar === secondChar.toUpperCase()) {
-				return { isClient: true, dependency: `JSX event handler: ${attrName}` };
-			}
-		}
-	}
-
-	// Check if statements for client detection pattern
-	if (node.type === AST_NODE_TYPES.IfStatement) {
-		if (isClientDetectionCondition(node.test)) {
-			return { isClient: true, dependency: "Client detection condition (typeof window check)" };
-		}
-	}
-
-	return { isClient: false };
-}
-
-// Export for testing - simplified import check
-export function checkImport(
-	node: TSESTree.Node,
-	options: {
-		allowlist: Record<string, unknown>;
-		isClientOnlyPackage: (pkg: string) => boolean;
-	}
-): { isClient: boolean; dependency?: string } {
-	if (node.type !== AST_NODE_TYPES.ImportDeclaration) {
-		return { isClient: false };
-	}
-
-	const importNode = node;
-	const moduleSpecifier = importNode.source.value;
-
-	// Skip completely allowlisted modules
-	if (options.allowlist[moduleSpecifier] === true) {
-		return { isClient: false };
-	}
-
-	// Check if it's a known client-only package
-	if (options.isClientOnlyPackage(moduleSpecifier)) {
-		return {
-			isClient: true,
-			dependency: `Import from client-only package: ${moduleSpecifier}`
-		};
-	}
-
-	// Skip type-only imports
-	if (importNode.importKind === "type") {
-		return { isClient: false };
-	}
-
-	// If we got here and there are named specifiers, we consider it potentially client code
-	if (importNode.specifiers.some((spec) => spec.type === AST_NODE_TYPES.ImportSpecifier)) {
-		return {
-			isClient: true,
-			dependency: `Potentially client import: ${moduleSpecifier}`
-		};
-	}
-
-	return { isClient: false };
-}
-
-// Expose internal function implementations for testing (these match the signatures of the exported functions)
-export function _checkNodeForClientAPIs(
-	node: TSESTree.Node,
-	options?: {
-		checker?: unknown;
-		clientOnlyHooksExtended?: string[];
-		serverSafeHooksExtended?: string[];
-		browserAPIs?: string[];
-	}
-): { isClient: boolean; dependency?: string } {
-	// This is a test-friendly implementation of the internal function
-	const clientOnlyHooksForTest = options?.clientOnlyHooksExtended ?? clientOnlyHooks;
-	const serverSafeHooksForTest = options?.serverSafeHooksExtended ?? serverSafeHooks;
-	const browserAPIsForTest = options?.browserAPIs ?? browserAPIs;
-
-	// Check for React hook usage
-	if (isHookCall(node)) {
-		const callExpr = node as TSESTree.CallExpression;
-		const callee = callExpr.callee;
-		const hookName = callee.type === AST_NODE_TYPES.Identifier ? callee.name : "";
-
-		// Check if it's a client-only hook
-		if (clientOnlyHooksForTest.includes(hookName)) {
-			return { isClient: true, dependency: `React hook: ${hookName}` };
-		}
-
-		// If it's a server-safe hook, we can ignore it
-		if (serverSafeHooksForTest.includes(hookName)) {
-			return { isClient: false };
-		}
-
-		// Unknown hooks should be treated conservatively
-		return { isClient: true, dependency: `Unknown hook call: ${hookName}` };
-	}
-
-	// Check for DOM globals
-	if (node.type === AST_NODE_TYPES.Identifier) {
-		const name = node.name;
-
-		if (browserAPIsForTest.includes(name)) {
-			return { isClient: true, dependency: `Browser API: ${name}` };
-		}
-	}
-
-	// Check for window/document property access
-	if (node.type === AST_NODE_TYPES.MemberExpression) {
-		if (node.object.type === AST_NODE_TYPES.Identifier) {
-			const objectName = node.object.name;
-			if (browserAPIsForTest.includes(objectName)) {
-				const propName = node.property.type === AST_NODE_TYPES.Identifier ? node.property.name : "?";
-				return { isClient: true, dependency: `Browser API: ${objectName}.${propName}` };
-			}
-		}
-	}
-
-	// Check for JSX event handlers
-	if (node.type === AST_NODE_TYPES.JSXAttribute) {
-		const attrName = node.name.type === AST_NODE_TYPES.JSXIdentifier ? node.name.name : "";
-		// Check if attribute is an event handler (starts with 'on' followed by uppercase letter)
-		if (attrName.startsWith("on") && attrName.length > 2 && attrName[2] === attrName[2].toUpperCase()) {
-			return { isClient: true, dependency: `JSX event handler: ${attrName}` };
-		}
-	}
-
-	// Check for if statements with client detection
-	if (node.type === AST_NODE_TYPES.IfStatement && isClientDetectionCondition(node.test)) {
-		return { isClient: true, dependency: "Client detection condition" };
-	}
-
-	return { isClient: false };
-}
-
-export function _checkDynamicImport(node: TSESTree.Node): { isClient: boolean; dependency?: string } {
-	// This is a test-friendly implementation for dynamic imports
-	if (node.type === AST_NODE_TYPES.ImportExpression) {
-		const source = node.source;
-
-		// Check literal source
-		if (source.type === AST_NODE_TYPES.Literal && typeof source.value === "string") {
-			const importPath = source.value;
-
-			// Check if it's a client-only package
-			const isClientPackage = clientOnlyPackages.some((pkg) => importPath === pkg || importPath.startsWith(`${pkg}/`));
-
-			if (isClientPackage) {
-				return { isClient: true, dependency: `Dynamic import of client package: ${importPath}` };
-			}
-		}
-	}
-
-	return { isClient: false };
-}
-
-export function _resolveModulePath(moduleSpecifier: string, _fromFile: string): string | null {
-	// Test-friendly implementation of module path resolution
-	try {
-		// Simple implementation - just for testing
-		if (moduleSpecifier.startsWith(".")) {
-			// Relative import - simulate resolution
-			return `resolved-${moduleSpecifier}`;
-		} else if (moduleSpecifier.startsWith("/")) {
-			// Absolute import - simulate resolution
-			return moduleSpecifier;
-		} else {
-			// Node module - simulate resolution
-			return `node_modules/${moduleSpecifier}`;
-		}
-	} catch (_error) {
-		return null;
-	}
-}
-
-export function _isNodeModule(path: string): boolean {
-	return path.includes("node_modules");
-}
-
-export function _analyzeModuleForClientDependencies(filePath: string, depth: number, visitedFiles?: Set<string>): boolean {
-	// Test-friendly implementation
-	// This is a simplified version just for testing
-	const visited = visitedFiles ?? new Set<string>();
-
-	if (visited.has(filePath)) {
-		return false; // Prevent infinite recursion
-	}
-
-	visited.add(filePath);
-
-	// First check depth limit before analyzing
-	if (depth <= 0) {
-		return false;
-	}
-
-	// Mock some client-side indicators for test files
-	if (filePath.includes("client") || filePath.includes("browser")) {
-		return true;
-	}
-
-	// Return false for tests
-	return false;
-}
-
-// Expose the visitor patterns from the rule's create function for testing
-export function _createVisitors(
-	_context: unknown, // Prefix with underscore to mark as intentionally unused
-	options: {
-		allowlist?: Record<string, unknown>;
-		traceDepth?: number;
-		traceDependencies?: boolean;
-		moduleCategories?: { clientModules: string[] };
-	} = {}
-): Record<string, (node: TSESTree.Node) => void> {
-	const _allowlist = options.allowlist ?? {};
-	const _traceDepth = options.traceDepth ?? 1;
-	const _traceDependencies = options.traceDependencies ?? true;
-	const _moduleCategories = options.moduleCategories ?? { clientModules: [] };
-
-	// Maintain state between visitor calls
-	let _hasUseClientDirective = false;
-	let _clientDependencies: string[] = [];
-	let _hasExports = false;
-	let _hasDirectClientAPIs = false;
-
-	return {
-		Program(node: TSESTree.Node): void {
-			// Do initial setup
-			_hasUseClientDirective = false;
-			_clientDependencies = [];
-			_hasExports = false;
-			_hasDirectClientAPIs = false;
-
-			// Check for 'use client' directive at top of file
-			const program = node as TSESTree.Program;
-			const comments = program.comments ?? [];
-			const directiveComments = comments.filter((comment) => {
-				const text = comment.value.trim();
-				return text === "use client";
-			});
-			_hasUseClientDirective = directiveComments.length > 0;
-		},
-
-		CallExpression(node: TSESTree.Node): void {
-			// Check for client-only hooks
-			const callExpression = node as TSESTree.CallExpression;
-			if (callExpression.callee.type === AST_NODE_TYPES.Identifier) {
-				const result = checkForClientAPIs(node);
-				if (result.isClient && result.dependency) {
-					_clientDependencies.push(result.dependency);
-					_hasDirectClientAPIs = true;
-				}
-			}
-		},
-
-		VariableDeclarator(node: TSESTree.Node): void {
-			// Check for window access
-			const declarator = node as TSESTree.VariableDeclarator;
-			if (declarator.init?.type === AST_NODE_TYPES.MemberExpression) {
-				if (declarator.init.object.type === AST_NODE_TYPES.Identifier && declarator.init.object.name === "window") {
-					_clientDependencies.push("Browser API: window");
-					_hasDirectClientAPIs = true;
-				}
-			}
-		},
-
-		ExpressionStatement(node: TSESTree.Node): void {
-			// Check for document access
-			const statement = node as TSESTree.ExpressionStatement;
-			if (statement.expression.type === AST_NODE_TYPES.MemberExpression) {
+			// Track variable declarations that alias DOM APIs
+			VariableDeclarator(node: TSESTree.VariableDeclarator): void {
 				if (
-					statement.expression.object.type === AST_NODE_TYPES.Identifier &&
-					statement.expression.object.name === "document"
+					node.init?.type === AST_NODE_TYPES.MemberExpression &&
+					node.init.object.type === AST_NODE_TYPES.Identifier &&
+					node.init.property.type === AST_NODE_TYPES.Identifier &&
+					node.id.type === AST_NODE_TYPES.Identifier
 				) {
-					_clientDependencies.push("Browser API: document");
-					_hasDirectClientAPIs = true;
+					const objectName = node.init.object.name;
+					const propertyName = node.init.property.name;
+					const importInfo = imports.get(objectName);
+
+					if (importInfo?.isNamespaceImport && DOM_APIS.has(propertyName)) {
+						domAliases.set(node.id.name, propertyName);
+					}
 				}
-			}
-		},
+			},
 
-		ImportDeclaration(node: TSESTree.Node): void {
-			// Check for client-only package imports
-			const importDecl = node as TSESTree.ImportDeclaration;
-			const source = importDecl.source.value;
-
-			// Check if the import is from a known client-only package
-			if (isKnownClientOnlyPackage(source, _moduleCategories)) {
-				_clientDependencies.push(`Client package: ${source}`);
-				_hasDirectClientAPIs = true;
-				return;
-			}
-
-			// Check if it's in the allowlist
-			if (_allowlist[source] === true) {
-				// Skip allowlisted modules
-				return;
-			}
-		},
-
-		ImportExpression(node: TSESTree.Node): void {
-			// Check for dynamic imports
-			const importExpr = node as TSESTree.ImportExpression;
-			const result = _checkDynamicImport(importExpr);
-			if (result.isClient && result.dependency) {
-				_clientDependencies.push(result.dependency);
-			}
-		},
-
-		ExportNamedDeclaration(): void {
-			_hasExports = true;
-		},
-
-		ExportDefaultDeclaration(): void {
-			_hasExports = true;
-		}
-	};
-}
+			Identifier: checkIdentifier,
+			MemberExpression: checkMemberExpression
+		};
+	}
+});
