@@ -5,7 +5,7 @@ import type { RuleContext, RuleListener } from "@typescript-eslint/utils/ts-esli
 /**
  * The type of message id to use.
  */
-export type MessageIds = "destructureReact";
+export type MessageIds = "destructureReact" | "removeUnusedReact";
 
 type RuleOptions = [];
 
@@ -23,15 +23,14 @@ function isReactImport(node: TSESTree.ImportDeclaration | TSESTree.VariableDecla
 }
 
 /**
- * Gets any type parameters associated with a React hook call
+ * Gets any type parameters associated with a React hook call or type reference
  */
 function getTypeParameters(
-	node: TSESTree.MemberExpression,
+	node: TSESTree.MemberExpression | TSESTree.TSTypeReference,
 	context: Readonly<RuleContext<MessageIds, RuleOptions>>
 ): string {
 	const parent = node.parent;
-	// This is a defensive check to ensure that we don't crash if the parent is not a TSTypeParameterInstantiation
-	/* v8 ignore next 8 */
+	/* v8 ignore start */
 	if (parent.type === AST_NODE_TYPES.TSTypeParameterInstantiation) {
 		try {
 			return context.sourceCode.getText(parent);
@@ -41,13 +40,30 @@ function getTypeParameters(
 		}
 	}
 	return "";
+	/* v8 ignore end */
+}
+
+/**
+ * Gets the full type reference text for a node
+ */
+function getFullTypeReferenceText(
+	node: TSESTree.TSTypeReference,
+	context: Readonly<RuleContext<MessageIds, RuleOptions>>
+): string {
+	/* v8 ignore start */
+	try {
+		return context.sourceCode.getText(node);
+	} catch {
+		return "";
+	}
+	/* v8 ignore end */
 }
 
 /**
  * The destructure-react-imports rule.
  * This rule is type-aware and will preserve TypeScript types in the output.
  */
-export const destructureReactImports = ESLintUtils.RuleCreator<{
+export const destructureReactImportsRule = ESLintUtils.RuleCreator<{
 	recommended: boolean;
 }>((name) => `https://spencerbeg.gs/rule/${name}`)<RuleOptions, MessageIds>({
 	name: "destructure-react-imports",
@@ -60,7 +76,8 @@ export const destructureReactImports = ESLintUtils.RuleCreator<{
 		fixable: "code",
 		schema: [],
 		messages: {
-			destructureReact: "React imports should be destructured"
+			destructureReact: "React imports should be destructured",
+			removeUnusedReact: "Unused React import should be removed"
 		}
 	},
 	defaultOptions: [],
@@ -68,9 +85,15 @@ export const destructureReactImports = ESLintUtils.RuleCreator<{
 		let defaultImportNode: TSESTree.ImportDefaultSpecifier | TSESTree.VariableDeclarator | null = null;
 		let importName: string | null = null;
 		const usedReactMembers = new Set<string>();
+		const existingNamedImports = new Set<string>();
 		const reactUsages: {
 			node: TSESTree.MemberExpression;
 			typeParams: string;
+		}[] = [];
+		const typeReferences: {
+			node: TSESTree.TSTypeReference;
+			typeName: string;
+			fullText: string;
 		}[] = [];
 
 		return {
@@ -78,6 +101,17 @@ export const destructureReactImports = ESLintUtils.RuleCreator<{
 				if (isReactImport(node.parent)) {
 					defaultImportNode = node;
 					importName = node.local.name;
+
+					// Collect any existing named imports
+					const parent = node.parent;
+					parent.specifiers.forEach((specifier) => {
+						if (specifier.type === AST_NODE_TYPES.ImportSpecifier) {
+							const imported = specifier.imported;
+							if (imported.type === AST_NODE_TYPES.Identifier) {
+								existingNamedImports.add(imported.name);
+							}
+						}
+					});
 				}
 			},
 			VariableDeclarator(node: TSESTree.VariableDeclarator) {
@@ -99,6 +133,18 @@ export const destructureReactImports = ESLintUtils.RuleCreator<{
 					reactUsages.push({ node, typeParams });
 				}
 			},
+			TSTypeReference(node: TSESTree.TSTypeReference) {
+				if (
+					node.typeName.type === AST_NODE_TYPES.TSQualifiedName &&
+					node.typeName.left.type === AST_NODE_TYPES.Identifier &&
+					importName !== null &&
+					node.typeName.left.name === importName
+				) {
+					usedReactMembers.add(node.typeName.right.name);
+					const fullText = getFullTypeReferenceText(node, context);
+					typeReferences.push({ node, typeName: node.typeName.right.name, fullText });
+				}
+			},
 			"Program:exit"() {
 				if (!defaultImportNode || !importName) {
 					return;
@@ -107,13 +153,41 @@ export const destructureReactImports = ESLintUtils.RuleCreator<{
 				const node = defaultImportNode;
 
 				// Report error for unused React import
-				if (usedReactMembers.size === 0) {
+				if (usedReactMembers.size === 0 && existingNamedImports.size === 0) {
 					context.report({
 						node,
-						messageId: "destructureReact",
+						messageId: "removeUnusedReact",
 						fix(fixer) {
 							const parent = node.type === AST_NODE_TYPES.ImportDefaultSpecifier ? node.parent : node.parent;
-							return fixer.removeRange([parent.range[0], parent.range[1]]);
+							const sourceCode = context.sourceCode;
+
+							// Prepare to capture the code that comes after the import
+							const program = sourceCode.ast;
+							let remainingCode = "";
+
+							// Skip the import node and include all other nodes
+							let foundImport = false;
+							for (const programNode of program.body) {
+								if (programNode === parent) {
+									foundImport = true;
+								} else if (foundImport) {
+									// Only include code after the import
+									remainingCode += sourceCode.getText(programNode);
+									if (programNode !== program.body[program.body.length - 1]) {
+										remainingCode += "\n";
+									}
+								}
+							}
+
+							// Create a fix that preserves the expected formatting
+							if (remainingCode) {
+								const startRange = parent.range[0];
+								const endRange = sourceCode.text.length;
+								return fixer.replaceTextRange([startRange, endRange], remainingCode);
+							} else {
+								// If there's no remaining code, just remove the import
+								return fixer.remove(parent);
+							}
 						}
 					});
 					return;
@@ -129,7 +203,8 @@ export const destructureReactImports = ESLintUtils.RuleCreator<{
 						const endRange = parent.range[1];
 
 						// Sort members for consistent output
-						const sortedMembers = Array.from(usedReactMembers).sort();
+						const allMembers = new Set([...existingNamedImports, ...usedReactMembers]);
+						const sortedMembers = Array.from(allMembers).sort();
 
 						// Create the appropriate import/require statement
 						if (node.type === AST_NODE_TYPES.ImportDefaultSpecifier) {
@@ -146,6 +221,15 @@ export const destructureReactImports = ESLintUtils.RuleCreator<{
 								const replacement = node.property.name + typeParams;
 								fixes.push(fixer.replaceText(node, replacement));
 							}
+						});
+
+						// Replace type references
+						typeReferences.forEach(({ node, typeName, fullText }) => {
+							if (importName === null) return;
+							const typeParamsMatch = new RegExp(`${importName}\\.${typeName}(.*)`).exec(fullText);
+							const typeParams = typeParamsMatch?.[1] ?? "";
+							const replacement = typeName + typeParams;
+							fixes.push(fixer.replaceText(node, replacement));
 						});
 
 						return fixes;
